Add missing unit tests for new conversion functions

+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/expression/function/scalar/convert/ToGeoShapeTests.java
@@ -50,7 +50,7 @@ public class ToGeoShapeTests extends AbstractScalarFunctionTestCase {
                 "ToGeoShapeFromGeoGridEvaluator[in=Attribute[channel=0], dataType=" + gridType + "]",
                 gridType,
                 DataType.GEO_SHAPE,
-                v -> EsqlDataTypeConverter.geoGridToShape(v, gridType),
+                v -> EsqlDataTypeConverter.geoGridToShape((long) v, gridType),
                 List.of()
             );
         }
